[ALPS02390480] M4U : 3G/4G mode
[Detail] addr + size = 0x100000000

[Solution] unsigned int --> unsigned long long

[Feature] Power On Issue

MTK-Commit-Id: 3f02cbe27f3d5a3742f43bdcbe30264fe1ef9612

CR-Id: ALPS02390480

+++ drivers/misc/mediatek/m4u/2.0/m4u_pgtable.h
@@ -143,15 +143,13 @@ static inline imu_pte_t *imu_largepage_start(imu_pte_t *pte)
 	return (imu_pte_t *) (round_down((unsigned long)pte, (16 * 4)));
 }
 
-static inline unsigned int m4u_calc_next_mva(unsigned int addr, unsigned int end, unsigned int size)
+static inline unsigned long long m4u_calc_next_mva(unsigned long long addr, unsigned long long end, unsigned int size)
 {
-	/* addr + size may equal 0x100000000 */
-	unsigned long long __boundary = ((unsigned long long)addr
-									+(unsigned long long)size)
-									&(~((unsigned long long)size-1));
-	unsigned long long min = min_t(unsigned long long, __boundary, (unsigned long long)end);
+	/* addr + size may equal 0x100000000*/
+	unsigned long long __boundary = (addr + (unsigned long long)size) & (~((unsigned long long)size-1));
+	unsigned long long min = min_t(unsigned long long, __boundary, end);
 
-	return (unsigned int)min;
+	return min;
 }
 
 #endif
